requirements.txt
hydra-core==1.1.1
icecream~=2.1.1
matplotlib~=3.5.0
numpy~=1.21.4
omegaconf~=2.1.1
pandas~=1.3.4
Pillow==8.2.0
scipy==1.7.0
seaborn~=0.11.2
svglib==1.1.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
torch~=1.10.0
torchvision==0.8.2
tqdm~=4.62.3
